feat(seer): Move the advanced seer settings to their own page
